Allows for card meta properties.
Allows for setting meta attributes in your SetDefinitions without needing a matching label.
Meta card properties have to start with an _underscore, or the game will print out a warning when
setting up the card labels

Added UT for card properties

Removed the need to define meta properties in a CardDefinitions list.
Now meta properties just need to start with an underscore.