Update QUICKSTART.md
Everything in card_labels needs an entry in `original_font_sizes` or else an error occurs. `CGFCardFront.gd` does this for all labels it defines, but if you extend it and add more in `_ready` then that runs after its definition so you need to specify the font size yourself.